+++ src/store/actions/index.js
@@ -1,0 +1,3 @@
+import actions from 'actions';
+
+export default actions;
